splitting L2TestHelpers
